pyproject.toml
[project]
name = "adguard-telegram-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot==21.6",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
    "apscheduler==3.10.4",
    "sqlalchemy==2.0.36",
]